feat: [36231] cafe_edit decomposition, zustand, ssr
cafe_edit decompose, zustandify

cafe_edit ssr initial data

fix: ssr title

fix: small bugs, code style
