+++ components/cafe_edit/CafeEdit_Modal_New.jsx
@@ -1,24 +1,35 @@
 "use client";
 
-import { memo, useState } from "react";
+import { memo, useEffect, useState } from "react";
 import useMyAlert from "@/src/hooks/useMyAlert";
 import { MyAlert, MySelect, MyTextInput } from "@/ui/elements";
-import { Button, Dialog, DialogActions, DialogContent, DialogTitle, Grid, IconButton } from "@mui/material";
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  Grid,
+  IconButton,
+} from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
+import useCafeEditModalsStore from "./useCafeEditModalsStore";
 
 const CafeEdit_Modal_New = (props) => {
   const { open, fullScreen } = props;
 
-  const [item, setItem] = useState(props.item);
+  const activePoint = useCafeEditModalsStore((s) => s.activePoint);
 
-  const { isAlert, showAlert, closeAlert, alertStatus, alertMessage } = useMyAlert();
+  const [item, setItem] = useState(activePoint);
 
   const changeItem = (key, event) => {
-    setItem({
-      ...item,
+    setItem((prev) => ({
+      ...prev,
       [key]: event.target.value,
-    });
+    }));
   };
+
+  const { isAlert, showAlert, closeAlert, alertStatus, alertMessage } = useMyAlert();
 
   const save = () => {
     if (!item?.city_id) {
@@ -29,20 +40,17 @@ const CafeEdit_Modal_New = (props) => {
       showAlert("Необходимо указать адрес");
       return;
     }
-    if (!item?.addr) {
-      showAlert("Необходимо указать адрес");
-      return;
-    }
-
     props.save(item);
     onClose();
   };
 
   const onClose = () => {
-    setItem(null);
+    useCafeEditModalsStore.setState({ activeItem: null });
     closeAlert();
     props.onClose();
   };
+
+  useEffect(() => setItem(activePoint), [activePoint])
 
   return (
     <>
