+++ src/createEmotionCache.js
@@ -1,0 +1,6 @@
+import createCache from "@emotion/cache";
+
+// key must be 'mui' to match MUI's default
+export default function createEmotionCache() {
+  return createCache({ key: "mui", prepend: true });
+}
