+++ src/hooks/useFullScreen.js
@@ -1,0 +1,6 @@
+import { useMediaQuery, useTheme } from "@mui/material";
+
+export default function useFullScreen(){
+  const theme = useTheme();
+  return useMediaQuery(theme.breakpoints.down("sm"));
+}
